// TileMap.cpp cast on values to prevent loss.

+++ src/TileMap.cpp
@@ -27,16 +27,16 @@ bool TileMap::load(const std::string& tileset, sf::Vector2u tileSize, const int*
             sf::Vertex* quad = &m_vertices[(i + j * width) * 4];
 
             // define its 4 corners
-            quad[0].position = sf::Vector2f(i * tileSize.x, j * tileSize.y);
-            quad[1].position = sf::Vector2f((i + 1) * tileSize.x, j * tileSize.y);
-            quad[2].position = sf::Vector2f((i + 1) * tileSize.x, (j + 1) * tileSize.y);
-            quad[3].position = sf::Vector2f(i * tileSize.x, (j + 1) * tileSize.y);
+            quad[0].position = sf::Vector2f(i * (float)tileSize.x, j * (float)tileSize.y);
+            quad[1].position = sf::Vector2f((i + 1) * (float)tileSize.x, j * (float)tileSize.y);
+            quad[2].position = sf::Vector2f((i + 1) * (float)tileSize.x, (j + 1) * (float)tileSize.y);
+            quad[3].position = sf::Vector2f(i * (float)tileSize.x, (j + 1) * (float)tileSize.y);
 
             // define its 4 texture coordinates
-            quad[0].texCoords = sf::Vector2f(tu * tileSize.x, tv * tileSize.y);
-            quad[1].texCoords = sf::Vector2f((tu + 1) * tileSize.x, tv * tileSize.y);
-            quad[2].texCoords = sf::Vector2f((tu + 1) * tileSize.x, (tv + 1) * tileSize.y);
-            quad[3].texCoords = sf::Vector2f(tu * tileSize.x, (tv + 1) * tileSize.y);
+            quad[0].texCoords = sf::Vector2f(tu * (float)tileSize.x, tv * (float)tileSize.y);
+            quad[1].texCoords = sf::Vector2f((tu + 1) * (float)tileSize.x, tv * (float)tileSize.y);
+            quad[2].texCoords = sf::Vector2f((tu + 1) * (float)tileSize.x, (tv + 1) * (float)tileSize.y);
+            quad[3].texCoords = sf::Vector2f(tu * (float)tileSize.x, (tv + 1) * (float)tileSize.y);
         }
 
     return true;
